refactor: rewrite modules/cjsToEsm.cjs in ECMAScript notation
